+++ packages/fluent/src/TypeOrmConnector/test/advanced/typeOrm.repository.ts
@@ -1,8 +1,9 @@
 import { TypeORMDataModel } from './typeOrm.entity'
 import { TypeOrmConnector } from '../../TypeOrmConnector'
+import { MemoryDataSource } from '../memoryDataSource'
 
 export class TypeOrmRepository extends TypeOrmConnector<TypeORMDataModel> {
   constructor(relations?: any) {
-    super(TypeORMDataModel, relations, 'runningTest')
+    super(TypeORMDataModel, MemoryDataSource, relations)
   }
 }