+++ packages/fluent/src/TypeOrmConnector/test/relations/car/car.repositoryTypeOrm.ts
@@ -2,10 +2,11 @@ import { CarsEntity } from './car.entity'
 import { CarsEntityIn } from './car.dto'
 import { TypeOrmConnector } from '../../../TypeOrmConnector'
 import { UserRepository } from '../user/user.repositoryTypeOrm'
+import { MemoryDataSource } from '../../memoryDataSource'
 
 export class CarsRepository extends TypeOrmConnector<CarsEntity, CarsEntityIn> {
   constructor(relationQuery?: any) {
-    super(CarsEntity, relationQuery, 'runningTest')
+    super(CarsEntity, MemoryDataSource, relationQuery)
   }
 
   public user = () => this.belongsTo<UserRepository>(UserRepository, 'user')