+++ packages/fluent/src/TypeOrmConnector/test/relations/roles/roles_users.repositoryTypeOrm.ts
@@ -1,12 +1,13 @@
 import { RoleUserEntityIn } from './roles_user.dto'
 import { RolesUser } from './roles_user.entity'
 import { TypeOrmConnector } from '../../../TypeOrmConnector'
+import { MemoryDataSource } from '../../memoryDataSource'
 
 export class RoleUsersRepository extends TypeOrmConnector<
   RolesUser,
   RoleUserEntityIn
 > {
   constructor(relations?: any) {
-    super(RolesUser, relations, 'runningTest')
+    super(RolesUser, MemoryDataSource, relations)
   }
 }