+++ packages/eslint/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @goatlab/eslint
 
+## 0.1.2
+
+### Patch Changes
+
+- Update TypeOrm base connector
+
 ## 0.1.1
 
 ### Patch Changes